% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSeal.R
\name{MiSeal}
\alias{MiSeal}
\title{Separate a minutiae pattern.}
\usage{
MiSeal(imagePath, roiPath, minutiaePath, T = 1e+06, burnIn = 10000)
}
\arguments{
\item{imagePath}{character string containing the path to the fingerprint image.
The image has to be a greyscale image, ideally binarised.}

\item{roiPath}{character string containing the path to the file containing the region of interest
The file should be a csv file of the same dimension as the used image. We use the convention that 1 denotes background (is not ROI) and 0 denotes foreground pixels (is ROI).}

\item{minutiaePath}{character string containing the path to the minutiae file.
 The minutiae template has to be a .txt file, written in the following fashion: \cr
image width \cr
image height \cr
image resolution \cr
number of minutiae \cr
x-coordinate1 y-coordinate1 orientation1 \cr
x-coordinate2 y-coordinate2 orientation2 \cr
x-coordinate3 y-coordinate3 orientation3 \cr
...  \cr}

\item{T}{number of iterations of the MCMC algorithm excluding burn in time.}

\item{burnIn}{number of burn in iterations.}
}
\value{
A list containing the traces of the parameters \eqn{(beta, gamma, lambda)}, the distribution of \eqn{W}, the acceptance rates, the marginal posteriors of \eqn{W}, the used MPLE and the considered point pattern as \code{spatstat ppp} with corresponding intensity image.
}
\description{
Separate a minutiae pattern.
}
\details{
Classify a given minutiae pattern into necessary and random minutiae. The minutiae pattern is assumed to be a superposition of two independent processes Xi and Eta where Xi is a homogeneous Poisson process and Eta is an inhomogeneous Strauss process with hard core.
}
\examples{
\donttest{
imagePath <- system.file("extdata", "FVC2002_DB2_IM_7_6_image.png", package = "MiSeal")
roiPath <- system.file("extdata", "FVC2002_DB2_IM_7_6_roi.csv", package = "MiSeal")
minutiaePath <- system.file("extdata", "FVC2002_DB2_IM_7_6_minutiae.txt", package = "MiSeal")
MiSealTest <- MiSeal( imagePath, roiPath, minutiaePath, T = 2000, burnIn = 1000)

dev.new()
par(mfrow = c(2,2), mar = c(4, 4, 2, 2))
plot((na.omit(MiSealTest$beta)), type = "l", xlab = "iteration", ylab = "beta", main = "trace plot for prefactor beta")
abline(h = MiSealTest$thetaHat[1], lty = "dotted", col = "chartreuse4", lwd = 2)

plot((na.omit(MiSealTest$gamma)), type = "l", xlab = "iteration", ylab = "gamma", main = "trace plot for interaction parameter gamma")
abline(h = MiSealTest$thetaHat[2], lty = "dotted", col = "chartreuse4", lwd = 2)

plot((na.omit(MiSealTest$lambda)), type = "l", xlab = "iteration", ylab = "lambda", main = "trace plot for noise intensity lambda")
}
barplot( MiSealTest$marginalW, names.arg = 1:length(MiSealTest$marginalW), main = "posterior probabilities for minutiae being necessary", col = 2, ylim = c(0,1) )
}
