% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSeal.R
\name{posteriorRatio}
\alias{posteriorRatio}
\title{Calculate ratio of posteriors.}
\usage{
posteriorRatio(
  theta,
  thetaNew,
  auxiliaryPattern,
  auxiliaryPatternDistances,
  W,
  window,
  nrep = 1e+06,
  intensityImage,
  thetaHat,
  distancesBetweenPoints,
  hc
)
}
\arguments{
\item{theta}{a vector containing the current parameters.}

\item{thetaNew}{a vector containing the new parameter.}

\item{auxiliaryPattern}{a point pattern (\code{spatstat ppp}) containing the current auxiliary point pattern corresponding to theta.}

\item{auxiliaryPatternDistances}{a matrix containing the pairwise distances between all points of the auxiliary point pattern.}

\item{W}{the current label vector.}

\item{window}{the region of interest window of the point pattern.}

\item{nrep}{the maximum number of interactions used for generating the new (proposed) auxiliary point pattern from the Strauss process.}

\item{intensityImage}{a pixel image (\code{spatstat im}) containing the intensity of necessary minutiae.}

\item{thetaHat}{the maximum pseudo likelihood estimate from the burn in.}

\item{distancesBetweenPoints}{a matrix containing the pairwise distances between all points of the minutiae pattern.}

\item{hc}{the hard core distance.}
}
\value{
Ratio of the posterior likelihoods for the proposed model at the current iteration.
}
\description{
Calculate ratio of posteriors.
}
