% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSeal.R
\name{qRatio}
\alias{qRatio}
\title{Compute the ratio of proposals.}
\usage{
qRatio(theta, thetaNew, type = "mv.lognormal")
}
\arguments{
\item{theta}{a vector containing the current parameters.}

\item{thetaNew}{a vector containing the new parameter.}

\item{type}{the proposal used. Yet only log-normal proposal implemented.}
}
\value{
Ratio of proposals.
}
\description{
Compute the ratio of proposals.
}
