% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSeal.R
\name{sR}
\alias{sR}
\title{Compute number of \eqn{r}-close pairs of points.}
\usage{
sR(r, W, distancesBetweenPoints)
}
\arguments{
\item{r}{the interaction distance of the Strauss process.}

\item{W}{the current label vector.}

\item{distancesBetweenPoints}{a matrix containing the pairwise distances between all points of the point pattern.}
}
\value{
Number of \eqn{r}-close pairs of points.
}
\description{
Compute number of \eqn{r}-close pairs of points.
}
