% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSeal.R
\name{thetaInitialise}
\alias{thetaInitialise}
\title{Initialise data for algorithm.}
\usage{
thetaInitialise(intensityPrior, interactionPrior, rPrior = r, noisePrior)
}
\arguments{
\item{intensityPrior}{a vector containing the parameters \eqn{a,b,m} for the prior of the intensity parameter beta (Gamma distribution with parameters \eqn{a / m}, \eqn{b / m}).}

\item{interactionPrior}{a vector containing the parameters \eqn{p,q} for the prior of the interaction parameter gamma (Beta distribution with parameters \eqn{a} and \eqn{b}).}

\item{rPrior}{the prior for the soft core distance. Default value is the pre-calculated value \eqn{r = 3 * hc}.}

\item{noisePrior}{a vector containing the parameters \eqn{a,b} for the prior for the intensity of the random minutiae (Gamma distribution with parameters \eqn{a} and \eqn{b}).}
}
\value{
A sample of the parameter vector \eqn{(beta, gamma, r, lambda)} from the prior.
}
\description{
Initialise data for algorithm.
}
\details{
Initialise data for algorithm. Convention for input of priors: first two components are parameters of gamma distribution, third is defines a multiplicative constant.
}
