% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSeal.R
\name{thetaUpdate}
\alias{thetaUpdate}
\title{Update parameter vector.}
\usage{
thetaUpdate(
  theta,
  auxiliaryPattern,
  auxiliaryPatternDistances,
  W,
  window,
  intensityPrior,
  interactionPrior,
  noisePrior,
  intensityImage,
  thetaHat,
  distancesBetweenPoints,
  hc,
  Sigma,
  nrep
)
}
\arguments{
\item{theta}{a vector containing the current parameters.}

\item{auxiliaryPattern}{a point pattern (\code{spatstat ppp}) containing the current auxiliary point pattern corresponding to theta.}

\item{auxiliaryPatternDistances}{a matrix containing the pairwise distances between all points of the auxiliary point pattern.}

\item{W}{the current label vector.}

\item{window}{the region of interest window of the point pattern.}

\item{intensityPrior}{a vector containing the parameters \eqn{a,b,m} for the prior of the intensity parameter beta (Gamma distribution with parameters \eqn{a / m}, \eqn{b / m}).}

\item{interactionPrior}{a vector containing the parameters \eqn{p,q} for the prior of the interaction parameter gamma (Beta distribution with parameters \eqn{a} and \eqn{b}).}

\item{noisePrior}{a vector containing the parameters \eqn{a,b} for the prior for the intensity of the random minutiae (Gamma distribution with parameters \eqn{a} and \eqn{b}).}

\item{intensityImage}{a pixel image (\code{spatstat im}) containing the intensity of necessary minutiae.}

\item{thetaHat}{the maximum pseudo likelihood estimate from the burn in.}

\item{distancesBetweenPoints}{a matrix containing the pairwise distances between all points of the minutiae pattern.}

\item{hc}{the hard core distance.}

\item{Sigma}{the covariance matrix of the parameters beta and gamma.}

\item{nrep}{the maximum number of interactions used for generating the new (proposed) auxiliary point pattern from the Strauss process.}
}
\value{
List containing the updated parameter vector (if discarded the old one), the corresponding auxiliary pattern, their pairwise distance matrix. Moreover, we return the index of the updated parameter, whether the proposal was accepted (1 for accept, 0 for reject) and the old auxiliary pattern.
}
\description{
Update parameter vector.
}
