% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSeal.R
\name{tR}
\alias{tR}
\title{Compute number of \eqn{r}-close neighbours of \eqn{z_i}.}
\usage{
tR(r, W, distancesBetweenPoints, index)
}
\arguments{
\item{r}{the interaction distance of the Strauss process.}

\item{W}{the current label vector.}

\item{distancesBetweenPoints}{a matrix containing the pairwise distances between all points of the point pattern.}

\item{index}{the index of the considered point \eqn{z_i}.}
}
\value{
Number of r-close neighbours of \eqn{z_i}.
}
\description{
Compute number of \eqn{r}-close neighbours of \eqn{z_i}.
}
