% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSeal.R
\name{wInitialise}
\alias{wInitialise}
\title{Initial assignment \eqn{W} of minutiae to the necessary and random sub-patterns.}
\usage{
wInitialise(n, distribution = "ber", pW)
}
\arguments{
\item{n}{length of the label vector}

\item{distribution}{distribution for drawing an initial label vector from. Yet only i.i.d. Bernoulli implemented.}

\item{pW}{the success probability of the Bernoulli trial.}
}
\value{
A \eqn{{0,1}} vector of length \eqn{n} as \eqn{n}-sample from a Bernoulli distribution.
}
\description{
Initial assignment \eqn{W} of minutiae to the necessary and random sub-patterns.
}
